+++ packages/config-manager/src/utils/generate-config-ts-def.js
