+++ packages/config-manager/src/utils/write-config-in-cwd.js
